dag schedule change and added load_mode for dim tables

+++ plugins/operators/load_dimension.py
@@ -13,13 +13,15 @@ class LoadDimensionOperator(BaseOperator):
                  redshift_conn_id = '',
                  table_name = '',
                  sql_load_query = '',
+                 load_mode = 'append',
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
 
         self.redshift_conn_id = redshift_conn_id
         self.table_name = table_name
-        self.sql_load_query = sql_load_query
+        self.sql_load_query = sql_load_query,
+        self.load_mode = load_mode
 
     def execute(self, context):
         ''' Loads data from staging tables to Dimension table '''
@@ -28,7 +30,11 @@ class LoadDimensionOperator(BaseOperator):
             logging.info(f"START: Loading Dimension Table '{self.table_name}' - Started Execution")
 
             redshift_hook = PostgresHook(postgres_conn_id = self.redshift_conn_id)
-            redshift_hook.run(f"TRUNCATE TABLE {self.table_name}")
+
+            if(self.load_mode == 'truncate'):
+                logging.info(f"Truncating Dimension Table '{self.table_name}'")
+                redshift_hook.run(f"TRUNCATE TABLE {self.table_name}")
+
             redshift_hook.run(f"INSERT INTO {self.table_name} {self.sql_load_query}")
 
             logging.info(f"SUCCESS: Loading Dimension Table '{self.table_name}'  - Finished Execution")
